Update lecture-3.md
code from line no. 20 does not have declaration of "x". Thus code on line no. 21 will give error of "Uncaught reference....". Either add declaration of 'x' as it is in above code snippet or just change the output of line no 21 from undefined to "Uncaught reference..."(As I suggested in this change request).